+++ lib/fae/options.rb
@@ -1,6 +1,7 @@
 require 'devise'
 require 'judge'
 require_relative 'validation_helper_collection'
+require_relative 'helper_texter'
 
 module Fae
   # configurable defaults