+++ spec/dummy/config/initializers/inflections.rb
@@ -11,6 +11,13 @@
 # end
 
 # These inflection rules are supported but not enabled by default:
-# ActiveSupport::Inflector.inflections(:en) do |inflect|
-#   inflect.acronym 'RESTful'
-# end
+ActiveSupport::Inflector.inflections(:en) do |inflect|
+  inflect.acronym 'LOL'
+  inflect.acronym 'PDF'
+  inflect.acronym 'SEO'
+  inflect.acronym 'CEO'
+  inflect.acronym 'SKU'
+  inflect.acronym 'URL'
+  inflect.acronym 'AVA'
+  inflect.acronym 'CTA'
+end